pyproject.toml

[project]
name = "nats-js-prom"
version = "0.1.0"
description = "NATS JetStream exporter for Prometheus"
authors = [
    {name = "Mikołaj Nowak", email = "12396461+m3nowak@users.noreply.github.com"},
]
dependencies = [
    "pydantic>=1.10.8",
    "pyhumps>=3.8.0",
    "aiofiles>=23.1.0",
    "pyyaml>=6.0",
    "click>=8.1.3",
    "litestar>=2.0.0a7",
    "uvicorn>=0.22.0",
    "prometheus-client>=0.17.0",
    "nats-py[nkeys]>=2.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
nats-js-prom = "nats_js_prom:__main__.main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.autopep8]
max_line_length = 120

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]
[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.269",
    "autopep8>=2.0.2",
]
[tool.pdm.build]
package-dir = "src"
